Authentification works and add styles to home, services and company views

+++ routes/web.php
@@ -26,15 +26,17 @@ Route::post('/recrutement', 'RecruitmentController@RecruitmentForm')->name('recr
 
 Route::get('/agences', 'AgenciesController@index')->name('agencies');
 
-Route::get('/admin', 'AdminController@index')->name('admin');
-Route::get('/admin/agency/create', 'AdminController@create')->name('agency.create');
-Route::post('/admin/agency', 'AdminController@store')->name('agency.store');
-Route::delete('/admin/agency/delete/{id}', 'AdminController@delete')->name('agency.delete');
-Route::get('/admin/agency/{id}/edit', 'AdminController@edit')->name('agency.edit');
-Route::patch('/admin/agency/{id}/updtate', 'AdminController@update')->name('agency.update');
+Route::get('/admin', 'AdminController@index')->name('admin')->middleware('auth');
+Route::get('/admin/agency/create', 'AdminController@create')->name('agency.create')->middleware('auth');
+Route::post('/admin/agency', 'AdminController@store')->name('agency.store')->middleware('auth');
+Route::delete('/admin/agency/delete/{id}', 'AdminController@delete')->name('agency.delete')->middleware('auth');
+Route::get('/admin/agency/{id}/edit', 'AdminController@edit')->name('agency.edit')->middleware('auth');
+Route::patch('/admin/agency/{id}/updtate', 'AdminController@update')->name('agency.update')->middleware('auth');
 
-Route::get('/login', 'LoginController@show')->name('login')->middleware('guest');
-Route::post('/login', 'LoginController@login')->name('login')->middleware('guest');
+Route::get('/login', 'Auth\LoginController@show')->name('login')->middleware('guest');
+Route::post('/login', 'Auth\LoginController@login')->name('login')->middleware('guest');
+Route::get('/logout', 'Auth\LoginController@logout')->name('logout');
 
-Auth::routes();
+
+Auth::routes(['register' => false]);
